fix: 패널티 페이지에서 컴포넌트로 변경
